server/requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
sqlparse==0.5.2
tomlkit==0.13.2
